+++ packages/sdk/src/domain.ts
@@ -11,6 +11,7 @@ import type { IGetBalance } from "./types/balances"
 import type { IGenerateTokenId } from "./types/nft/generate-token-id"
 import type { IDeploy } from "./types/nft/deploy/domain"
 import type { IRestrictionSdk } from "./types/nft/restriction/domain"
+import type { IPreprocessMeta } from "./types/nft/mint/preprocess-meta"
 
 export interface IRaribleSdk {
 	apis: IApisSdk
@@ -31,6 +32,7 @@ export interface IApisSdk {
 
 export interface INftSdk {
 	transfer: ITransfer
+	preprocessMeta: IPreprocessMeta
 	mint: IMint
 	mintAndSell: IMintAndSell
 	burn: IBurn