+++ packages/sdk/src/sdk-blockchains/flow/index.ts
@@ -21,14 +21,16 @@ export function createFlowSdk(
 ): IRaribleInternalSdk {
 	const sdk = createFlowSdkInstance(wallet?.fcl, network, auth)
 	const sellService = new FlowSell(sdk, apis)
+	const mintService = new FlowMint(sdk, apis)
 
 	return {
 		nft: {
-			mint: new FlowMint(sdk, apis).prepare,
+			mint: mintService.prepare,
 			burn: new FlowBurn(sdk).burn,
 			transfer: new FlowTransfer(sdk).transfer,
 			generateTokenId: () => Promise.resolve(undefined),
 			deploy: nonImplementedAction,
+			preprocessMeta: mintService.preprocessMeta,
 		},
 		order: {
 			sell: sellService.sell,