+++ packages/sdk/src/sdk-blockchains/tezos/index.ts
@@ -1,8 +1,6 @@
 import type { TezosWallet } from "@rarible/sdk-wallet"
 import type { Maybe } from "@rarible/types/build/maybe"
-import { notImplemented } from "../../common/not-implemented"
-import type { IRaribleInternalSdk } from "../../domain"
-import type { IApisSdk } from "../../domain"
+import type { IApisSdk, IRaribleInternalSdk } from "../../domain"
 import { TezosSell } from "./sell"
 import { TezosFill } from "./fill"
 import { TezosBid } from "./bid"
@@ -15,7 +13,7 @@ import { TezosTokenId } from "./token-id"
 import { TezosCancel } from "./cancel"
 import { TezosBalance } from "./balance"
 import { TezosDeploy } from "./deploy"
-import { canTransfer } from "./restriction/can-transfer"
+import { canTransfer } from "./restriction"
 
 export function createTezosSdk(
 	wallet: Maybe<TezosWallet>,
@@ -25,21 +23,24 @@ export function createTezosSdk(
 	const apis = getTezosAPIs(network)
 	const maybeProvider = getMaybeTezosProvider(wallet?.provider, network)
 	const sellService = new TezosSell(maybeProvider, apis)
+	const mintService = new TezosMint(maybeProvider, apis)
+	const bidService = new TezosBid(maybeProvider, apis)
 
 	return {
 		nft: {
-			mint: new TezosMint(maybeProvider, apis).mint,
+			mint: mintService.mint,
 			burn: new TezosBurn(maybeProvider, apis).burn,
 			transfer: new TezosTransfer(maybeProvider, apis).transfer,
 			generateTokenId: new TezosTokenId(maybeProvider, apis).generateTokenId,
 			deploy: new TezosDeploy(maybeProvider, apis).deployToken,
+			preprocessMeta: mintService.preprocessMeta,
 		},
 		order: {
 			fill: new TezosFill(maybeProvider, apis).fill,
 			sell: sellService.sell,
 			sellUpdate: sellService.update,
-			bid: new TezosBid(maybeProvider, apis).bid,
-			bidUpdate: notImplemented,
+			bid: bidService.bid,
+			bidUpdate: bidService.update,
 			cancel: new TezosCancel(maybeProvider, apis).cancel,
 		},
 		balances: {
@@ -47,6 +48,7 @@ export function createTezosSdk(
 		},
 		restriction: {
 			canTransfer,
+
 		},
 	}
 }