+++ packages/sdk/src/sdk-blockchains/tezos/restriction/can-transfer.ts
@@ -4,23 +4,14 @@ import { convertUnionAddress } from "../common"
 
 const url = "https://hangzhounet.smartpy.io/chains/main/blocks/head/helpers/scripts/run_view"
 
-const tokenMapping = {
-	"KT1GXE3DGqyxTsrh6mHkfPtd9TFoGnK8vDv8": "KT1JfpeUECSWCNgCxdwof2nntxq8ua9swiR2",
-}
-
 export async function canTransfer(
 	itemId: ItemId, from: UnionAddress, to: UnionAddress,
 ): Promise<CanTransferResult> {
 	const parsed = itemId.split(":")
-	const contract = parsed[1]
 	const tokenId = parsed[2]
-	const whitelistContract = (tokenMapping as any)[contract]
-	if (whitelistContract === undefined) {
-		return { success: true }
-	}
 	const body = {
 		"chain_id": "NetXZSsxBpMQeAT",
-		"contract": whitelistContract,
+		"contract": "KT1J3Zzsz3cGYbeLrfvZBTQyuMepSuTspgon",
 		"entrypoint": "can_transfer",
 		"gas": "100000",
 		"input": {
@@ -36,10 +27,11 @@ export async function canTransfer(
 				},
 			],
 		},
-		"payer": "tz1hyc1CRQpjskJUUaGrh85UZXPi6kU4JuGd",
-		"source": "tz1hyc1CRQpjskJUUaGrh85UZXPi6kU4JuGd",
+		"payer": "KT1AguExF32Z9UEKzD5nuixNmqrNs1jBKPT8",
+		"source": "KT1AguExF32Z9UEKzD5nuixNmqrNs1jBKPT8",
 		"unparsing_mode": "Readable",
 	}
+	console.log("body is", JSON.stringify(body))
 	const response = await window.fetch(url, {
 		method: "POST",
 		headers: {
@@ -49,11 +41,15 @@ export async function canTransfer(
 		body: JSON.stringify(body),
 	})
 	const result: CheckResponse = await response.json()
+	console.log("result is", JSON.stringify(result))
 	if (result.data.string === "") {
 		return { success: true }
 	}
-	return { success: false, reason: "You are not authorized to purchase this Item" }
+	return { success: false, reason }
 }
+
+const reason = "Ubisoft Quartz NFTs are only available to Ubisoft players.\n" +
+  "Please read [Ubisoft Quartz’s FAQ](https://quartz.ubisoft.com/faq/) for more information."
 
 type ERROR_CODE = "FROM_RESTRICTED" | "TO_RESTRICTED" | "TO_NOT_ALLOWED" | "BAD_TOKEN_ID" | "ARCHETYPE_QUOTA_REACHED"
 | "ARCHOWNER_NOT_SET" | "ARCHLEDGER_NOT_SET" | "WHITELIST_ERROR"