+++ packages/sdk/src/types/nft/restriction/domain.ts
@@ -1,4 +1,4 @@
-import type { ItemId, UnionAddress } from "@rarible/types"
+import type { ContractAddress, ItemId, UnionAddress } from "@rarible/types"
 
 export type CanTransferResult = {
 	success: true
@@ -7,8 +7,40 @@ export type CanTransferResult = {
 	reason: string
 }
 
+export enum ItemActionType {
+	BURN = "BURN",
+	TRANSFER = "TRANSFER",
+	SELL = "SELL",
+	BID = "BID",
+}
+export enum CollectionActionType {
+	DEPLOY = "DEPLOY",
+	MINT = "MINT",
+}
+// export enum OrderActionType {
+//CANCEL = "CANCEL",
+//FILL = "FILL",
+// }
+
+
 export type IRestrictionSdk = {
-	canTransfer: (
-		itemId: ItemId, from: UnionAddress, to: UnionAddress,
-	) => Promise<CanTransferResult>
+	items: {
+		//get all common items actions
+		getActions: () => ItemActionType[],
+		//particular actions with arguments
+		canTransfer: (
+			itemId: ItemId, from: UnionAddress, to: UnionAddress,
+		) => Promise<CanTransferResult>
+		// canBurn,
+	}
+	collections: {
+		//get all common collection actions
+		getActions: () => CollectionActionType[],
+		canMint: (collectionId: ContractAddress) => boolean
+	}
+	// orders: {
+	// getActions: () => OrderActionType[],
+	// canCancel: () => boolean
+	// canFill: () => boolean
+	// }
 }