+++ packages/wallet/src/index.ts
@@ -72,10 +72,10 @@ export class TezosWallet implements AbstractWallet {
 		if (publicKey === undefined) {
 			throw new Error("Public key undefined")
 		}
-		const result = await sign(this.provider, message)
+		const result = await sign(this.provider, message, "message")
 		return {
 			signature: result.signature,
-			publicKey: result.edpk,
+			publicKey: `${result.edpk}_${result.prefix}`,
 		}
 	}
 }